Game rankings: Special case CHI Play proceedings
Starting in 2021, the main conference proceedings of CHI Play are published as
a special issue of journals/pacmhci, specifically the special issue that DBLP
codes as having issue number "CHI" (the DBLP website displays it as "CHI PLAY",
but the XML dump seems to shorten this to "CHI"). For our purposes we want to
treat this issue as a conf/chiplay proceedings.